pyproject.toml
[tool.poetry]
name = "sargilo"
version = "0.1.0"
description = "Dynamic loader for test data"
authors = ["Nick Lehmann <nicklehmann@protonmail.ch>"]

[tool.poetry.dependencies]
python = ">2.7"
"ruamel.yaml" = "*"
south = "*"

[tool.poetry.dev-dependencies]
django = "~1.4"

[tool.poetry.extras]
docs = ["sphinx", "recommonmark", "sphinx_rtd_theme"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
